restructure error module
